update to 64 bits types + a callbacks per relation type

+++ include/provenancefilter.h
@@ -23,56 +23,56 @@
 * @filter pointer to contain filter to read
 * read the current state of the node filter.
 */
-int provenance_get_node_filter( uint32_t* filter );
+int provenance_get_node_filter( uint64_t* filter );
 
 /*
 * @filter value of node filter
 * set node provenance capture filter.
 */
-int provenance_add_node_filter( uint32_t filter );
-int provenance_remove_node_filter( uint32_t filter );
+int provenance_add_node_filter( uint64_t filter );
+int provenance_remove_node_filter( uint64_t filter );
 int provenance_reset_node_filter( void );
 
 /*
 * @filter pointer to contain filter to read
 * read the current state of the node filter.
 */
-int provenance_get_propagate_node_filter( uint32_t* filter );
+int provenance_get_propagate_node_filter( uint64_t* filter );
 
 /*
 * @filter value of node filter
 * set node provenance propagate filter.
 */
-int provenance_add_propagate_node_filter( uint32_t filter );
-int provenance_remove_propagate_node_filter( uint32_t filter );
+int provenance_add_propagate_node_filter( uint64_t filter );
+int provenance_remove_propagate_node_filter( uint64_t filter );
 int provenance_reset_propagate_node_filter( void );
 
 /*
 * @filter pointer to contain filter to read
 * read the current state of the relation filter.
 */
-int provenance_get_relation_filter( uint32_t* filter );
+int provenance_get_relation_filter( uint64_t* filter );
 
 /*
 * @filter value of node filter
 * set relation provenance capture filter.
 */
-int provenance_add_relation_filter( uint32_t filter );
-int provenance_remove_relation_filter( uint32_t filter );
+int provenance_add_relation_filter( uint64_t filter );
+int provenance_remove_relation_filter( uint64_t filter );
 int provenance_reset_relation_filter( void );
 
 /*
 * @filter pointer to contain filter to read
 * read the current state of the relation filter.
 */
-int provenance_get_propagate_relation_filter( uint32_t* filter );
+int provenance_get_propagate_relation_filter( uint64_t* filter );
 
 /*
 * @filter value of node filter
 * set relation provenance propagate filter.
 */
-int provenance_add_propagate_relation_filter( uint32_t filter );
-int provenance_remove_propagate_relation_filter( uint32_t filter );
+int provenance_add_propagate_relation_filter( uint64_t filter );
+int provenance_remove_propagate_relation_filter( uint64_t filter );
 int provenance_reset_propagate_relation_filter( void );
 
 #endif
